+++ packages/vscode-avedon/CHANGELOG.md
@@ -1,5 +1,13 @@
 # avedon-vscode
 
+## 2.0.1
+
+### Patch Changes
+
+- Lockstep with publishable packages (same version train).
+- Updated dependencies
+  - @avedon/language-server@2.0.1
+
 ## 0.1.1
 
 ### Patch Changes